Update README.md
This is the earliest steps of the configuring this classroom site. For now the admin password is master123. thanks it worked well.
Remember I have not yet configured the admin panel hence the admin login page. 
By now we're successfully manage to deal with the user such that If the user entered the inputs credentials that are not in the database then He'll be told that incorrect.
  + For now I successfully configure the SQL injections and XSS vulnerabilities
  + Fun fact is that I didn't use the PDO but the vulnerabilities were controlled successfully